Moved function definitions to .cpp files.
